refactor(edge): move key related functionality to key
