+++ edge/edge.go
@@ -7,9 +7,7 @@ import (
 
 	"github.com/portainer/agent"
 	"github.com/portainer/agent/exec"
-	"github.com/portainer/agent/filesystem"
 	"github.com/portainer/agent/http"
-	"github.com/portainer/agent/http/client"
 	"github.com/portainer/agent/http/edgestacks"
 	"github.com/portainer/agent/http/tunnel"
 )
@@ -73,7 +71,7 @@ func (manager *EdgeManager) Enable(edgeKey string) error {
 	if edgeKey != "" {
 		log.Println("[DEBUG] [main,edge] [message: Edge key found in environment. Associating Edge key to cluster.]")
 
-		err := manager.associateEdgeKey(edgeKey)
+		err := manager.SetKey(edgeKey)
 		if err != nil {
 			return err
 		}
@@ -90,92 +88,6 @@ func (manager *EdgeManager) Enable(edgeKey string) error {
 
 func (manager *EdgeManager) ResetActivityTimer() {
 	manager.tunnelOperator.ResetActivityTimer()
-}
-
-func (manager *EdgeManager) retrieveEdgeKey(edgeKey string) (string, error) {
-
-	if edgeKey != "" {
-		log.Println("[INFO] [main,edge] [message: Edge key loaded from options]")
-		return edgeKey, nil
-	}
-
-	var keyRetrievalError error
-
-	edgeKey, keyRetrievalError = manager.retrieveEdgeKeyFromFilesystem()
-	if keyRetrievalError != nil {
-		return "", keyRetrievalError
-	}
-
-	if edgeKey == "" && manager.clusterService != nil {
-		edgeKey, keyRetrievalError = retrieveEdgeKeyFromCluster(manager.clusterService)
-		if keyRetrievalError != nil {
-			return "", keyRetrievalError
-		}
-	}
-
-	return edgeKey, nil
-}
-
-func (manager *EdgeManager) retrieveEdgeKeyFromFilesystem() (string, error) {
-	var edgeKey string
-
-	edgeKeyFilePath := fmt.Sprintf("%s/%s", agent.DataDirectory, agent.EdgeKeyFile)
-
-	keyFileExists, err := filesystem.FileExists(edgeKeyFilePath)
-	if err != nil {
-		return "", err
-	}
-
-	if keyFileExists {
-		filesystemKey, err := filesystem.ReadFromFile(edgeKeyFilePath)
-		if err != nil {
-			return "", err
-		}
-
-		log.Println("[INFO] [main,edge] [message: Edge key loaded from the filesystem]")
-		edgeKey = string(filesystemKey)
-	}
-
-	return edgeKey, nil
-}
-
-func retrieveEdgeKeyFromCluster(clusterService agent.ClusterService) (string, error) {
-	var edgeKey string
-
-	member := clusterService.GetMemberWithEdgeKeySet()
-	if member != nil {
-		httpCli := client.NewAPIClient()
-
-		memberAddr := fmt.Sprintf("%s:%s", member.IPAddress, member.Port)
-		memberKey, err := httpCli.GetEdgeKey(memberAddr)
-		if err != nil {
-			log.Printf("[ERROR] [main,edge,http,cluster] [message: Unable to retrieve Edge key from cluster member] [error: %s]", err)
-			return "", err
-		}
-
-		log.Println("[INFO] [main,edge] [message: Edge key loaded from cluster]")
-		edgeKey = memberKey
-	}
-
-	return edgeKey, nil
-}
-
-func (manager *EdgeManager) associateEdgeKey(edgeKey string) error {
-	err := manager.SetKey(edgeKey)
-	if err != nil {
-		return err
-	}
-
-	if manager.clusterService != nil {
-		tags := manager.clusterService.GetTags()
-		tags[agent.MemberTagEdgeKeySet] = "set"
-		err = manager.clusterService.UpdateTags(tags)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
 
 func (manager *EdgeManager) serveEdgeUI() {
