+++ http/handler/key/handler.go
@@ -15,7 +15,6 @@ type Handler struct {
 	*mux.Router
 	tunnelOperator agent.TunnelOperator
 	edgeManager    agent.EdgeManager
-	clusterService agent.ClusterService
 	edgeMode       bool
 }
 
@@ -23,12 +22,11 @@ type Handler struct {
 // It sets the associated handle functions for all the Edge key related HTTP endpoints.
 // This handler is meant to be used when the agent is started in Edge mode, all the API endpoints will return
 // a HTTP 503 service not available if edge mode is disabled.
-func NewHandler(tunnelOperator agent.TunnelOperator, clusterService agent.ClusterService, notaryService *security.NotaryService, edgeManager agent.EdgeManager, edgeMode bool) *Handler {
+func NewHandler(tunnelOperator agent.TunnelOperator, notaryService *security.NotaryService, edgeManager agent.EdgeManager, edgeMode bool) *Handler {
 	h := &Handler{
 		Router:         mux.NewRouter(),
 		edgeManager:    edgeManager,
 		tunnelOperator: tunnelOperator,
-		clusterService: clusterService,
 		edgeMode:       edgeMode,
 	}
